refactor: clubbed all leftsidebar menu opovers into single component
